refactored dashboard to be on two seperate files. fiished the proteges dashboard functionality
